requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Image Processing
Pillow==10.0.0

# Utilities
python-dotenv==1.0.0

# Optional: For production deployment
gunicorn==21.2.0

absl-py                      2.3.0
aiohappyeyeballs             2.6.1
aiohttp                      3.12.15
aiosignal                    1.4.0
anyio                        4.9.0
argon2-cffi                  25.1.0
argon2-cffi-bindings         21.2.0
arrow                        1.3.0
asttokens                    2.4.1
astunparse                   1.6.3
async-lru                    2.0.5
async-timeout                5.0.1
attrs                        25.3.0
babel                        2.17.0
base                         0.0.0
beautifulsoup4               4.13.4
bleach                       6.2.0
blinker                      1.9.0
bokeh                        2.4.3
cachetools                   5.5.2
certifi                      2025.6.15
cffi                         1.17.1
charset-normalizer           3.4.2
click                        8.1.8
clip                         1.0
cloudpickle                  3.1.1
colorama                     0.4.6
comm                         0.2.2
contourpy                    1.2.1
cvxopt                       1.3.0
cycler                       0.12.1
dask                         2021.12.0
debugpy                      1.8.14
decorator                    5.2.1
defusedxml                   0.7.1
et_xmlfile                   2.0.0
exceptiongroup               1.3.0
executing                    0.8.3
fastjsonschema               2.21.1
Flask                        2.3.3
Flask-Cors                   3.0.10
flaskr                       1.0.0          E:\flask\examples\tutorial
flatbuffers                  25.2.10
fonttools                    4.58.4
fqdn                         1.5.1
frozenlist                   1.7.0
fsspec                       2025.7.0
ftfy                         6.3.1
gast                         0.4.0
google-auth                  2.40.3
google-auth-oauthlib         0.4.6
google-pasta                 0.2.0
greenlet                     3.2.4
grpcio                       1.73.1
h11                          0.16.0
h5py                         3.14.0
httpcore                     1.0.9
httpx                        0.28.1
idna                         3.10
imageio                      2.37.0
imagenetv2_pytorch           0.1
imbalanced-learn             0.10.1
importlib_metadata           8.7.0
ipykernel                    6.29.5
ipython                      8.18.1
ipython-genutils             0.2.0
ipywidgets                   7.8.5
isoduration                  20.11.0
itsdangerous                 2.2.0
jedi                         0.19.2
Jinja2                       3.1.6
joblib                       1.5.1
json5                        0.12.0
jsonpath-ng                  1.7.0
jsonpointer                  3.0.0
jsonschema                   4.24.0
jsonschema-specifications    2025.4.1
jupyter                      1.1.1
jupyter_client               8.6.3
jupyter-console              6.6.3
jupyter_core                 5.8.1
jupyter-events               0.12.0
jupyter-lsp                  2.2.5
jupyter_server               2.16.0
jupyter_server_terminals     0.5.3
jupyterlab                   4.4.4
jupyterlab_pygments          0.3.0
jupyterlab_server            2.27.3
jupyterlab_widgets           1.1.11
keras                        2.10.0
Keras-Preprocessing          1.1.2
kiwisolver                   1.4.7
libclang                     18.1.1
llvmlite                     0.43.0
locket                       1.0.0
Markdown                     3.8.2
MarkupSafe                   3.0.2
matplotlib                   3.6.3
matplotlib-inline            0.1.7
Metaphone                    0.6
mistune                      3.1.3
multidict                    6.6.4
nbclient                     0.10.2
nbconvert                    7.16.6
nbformat                     5.10.4
nest-asyncio                 1.6.0
networkx                     3.2.1
nltk                         3.9.1
notebook                     7.4.3
notebook_shim                0.2.4
numba                        0.60.0
numpy                        1.23.5
oauthlib                     3.3.1
opencv-contrib-python        4.6.0.66
opencv-python                4.6.0.66
openpyxl                     3.1.2
opt_einsum                   3.4.0
overrides                    7.7.0
packaging                    25.0
pandas                       1.5.3
pandocfilters                1.5.1
parso                        0.8.4
partd                        1.4.2
pillow                       11.2.1
pip                          25.1
platformdirs                 4.3.8
ply                          3.11
prometheus_client            0.22.1
prompt_toolkit               3.0.51
propcache                    0.3.2
protobuf                     3.19.6
psutil                       7.0.0
pure_eval                    0.2.3
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pycparser                    2.22
pydantic                     1.10.22
Pygments                     2.19.2
pyod                         1.0.9
pyparsing                    3.2.3
pystache                     0.6.8
python-crfsuite              0.9.11
python-dateutil              2.9.0.post0
python-json-logger           3.3.0
python-Levenshtein           0.12.2
python-stdnum                1.20
pytz                         2025.2
PyWavelets                   1.6.0
pywin32                      310
pywinpty                     2.0.15
PyYAML                       6.0.2
pyzmq                        27.0.0
referencing                  0.36.2
regex                        2021.11.10
requests                     2.32.4
requests-oauthlib            2.0.0
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rpds-py                      0.25.1
rsa                          4.9.1
scikit-image                 0.19.3
scikit-learn                 1.1.3
scipy                        1.9.3
seaborn                      0.12.2
Send2Trash                   1.8.3
setuptools                   78.1.1
six                          1.17.0
sniffio                      1.3.1
soupsieve                    2.7
SQLAlchemy                   1.4.54
stack-data                   0.5.1
tabulate                     0.8.7
tensorboard                  2.10.1
tensorboard-data-server      0.6.1
tensorboard-plugin-wit       1.8.1
tensorflow                   2.10.0
tensorflow-estimator         2.10.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    3.1.0
terminado                    0.18.1
threadpoolctl                3.6.0
tifffile                     2023.7.10
tinycss2                     1.4.0
tomli                        2.2.1
toolz                        1.0.0
torch                        1.13.1+cu117
torchaudio                   0.13.1+cu117
torchvision                  0.14.1+cu117
tornado                      6.5.1
tqdm                         4.67.1
traitlets                    5.14.3
types-python-dateutil        2.9.0.20250516
typing_extensions            4.14.0
uri-template                 1.3.0
urllib3                      2.5.0
varname                      0.8.3
wcwidth                      0.2.13
webcolors                    24.11.1
webencodings                 0.5.1
websocket-client             1.8.0
Werkzeug                     3.1.3
wheel                        0.45.1
widgetsnbextension           3.6.10
wordcloud                    1.9.4
wrapt                        1.17.2
xgboost                      1.7.6
yarl                         1.20.1
zipp                         3.23.0